Added TV-B-Gone code. Added enclosure file.
I added the TV-B-Gone off codes to the main file. Works fairly well.
Distance is only 5-10 feet. It takes 20-25 seconds to broadcast all the
IR codes.

Added enclosure file. Enclosure was modified to have a spot for the IR
LED on layer 3. Layer 2 was increased in thickness by 1.5mm to allow for
more solder joint height. Old versions required users to cut the solder
joints on all parts very close to the board. Hard for beginners.